f: minor fixes and detail for collectible on both views
